React: Simplify docs preset
