+++ app/react/src/server/preset.ts
@@ -9,6 +9,5 @@ export const config: StorybookConfig['config'] = (entries = []) => [
 export const addons: StorybookConfig['addons'] = [
   require.resolve('./framework-preset-react'),
   require.resolve('./framework-preset-cra'),
-  require.resolve('./framework-preset-react-docgen'),
   require.resolve('./framework-preset-react-docs'),
 ];
